Recommend user to not use dots in AWS bucket names
This is for both compatibility and performance reasons, and a mistake we
made ourselves multiple times. Fog warns you about the performance if
you decide to do it anyways:

```
[fog][WARNING] fog: the specified s3 bucket name(cfapps.io-cc-buildpacks)
contains a '.' so is not accessible over https as a virtual hosted bucket,
which will negatively impact performance. For details see:
http://docs.amazonwebservices.com/AmazonS3/latest/dev/BucketRestrictions.html
```

Related reads:
- https://github.com/cloudfoundry/cloud_controller_ng/issues/175#issuecomment-36576520
- https://github.com/cloudfoundry/cf-release/pull/573
- https://groups.google.com/a/cloudfoundry.org/forum/#!topic/vcap-dev/h4ZwbAQDyEw
- https://groups.google.com/a/cloudfoundry.org/forum/#!topic/bosh-users/0waAA386RuI
- https://github.com/cloudfoundry/cf-release/pull/312